feat: Use a specific buildpack version to lock Node versions
